renamed formatStarkKey to formatStarkPublicKey
