+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "dvf-client-js",
-  "version": "0.1.8",
+  "version": "0.1.9",
   "main": "src/dvf.js",
   "contributors": [
     "Henrique Matias <hems.inlet@gmail.com>",
