+++ src/api/register.test.js
@@ -33,7 +33,7 @@ describe('dvf.register', () => {
       y: fullPublicKey.pub.getY().toString('hex')
     }
 
-    const starkPublicKey = dvf.stark.formatStarkKey(tempKey)
+    const starkPublicKey = dvf.stark.formatStarkPublicKey(tempKey)
     
     nock(dvf.config.api)
       .post('/v1/trading/w/preRegister', body => {
