+++ src/api/walletConnect/getPublicKey.js
@@ -13,5 +13,5 @@ module.exports = async (dvf, path) => {
     y: tempKey.substr(66)
   }
   await transport.close()
-  return dvf.stark.formatStarkKey(starkPublicKey)
+  return dvf.stark.formatStarkPublicKey(starkPublicKey)
 }
