+++ src/lib/stark/formatStarkPublicKey.js
